+++ shared/src/commonMain/kotlin/site/remlit/snowdrop/util/Snackbar.kt
@@ -3,6 +3,6 @@ package site.remlit.snowdrop.util
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.runtime.staticCompositionLocalOf
 
-val SnackbarController = staticCompositionLocalOf<SnackbarHostState> {
+val LocalSnackbarController = staticCompositionLocalOf<SnackbarHostState> {
 	error("SnackbarHostState not found")
 }