+++ shared/src/commonMain/kotlin/site/remlit/snowdrop/view/StatusInteractionDetailView.kt
@@ -1,14 +1,8 @@
 package site.remlit.snowdrop.view
 
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.PrimaryScrollableTabRow
 import androidx.compose.material3.Tab
 import androidx.compose.material3.Text
@@ -19,16 +13,12 @@ import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.dp
 import org.jetbrains.compose.resources.painterResource
 import org.jetbrains.compose.resources.stringResource
 import site.remlit.snowdrop.api.statuses.getStatusFavouritedBy
 import site.remlit.snowdrop.api.statuses.getStatusReactions
 import site.remlit.snowdrop.api.statuses.getStatusRebloggedBy
-import site.remlit.snowdrop.component.Avatar
+import site.remlit.snowdrop.component.AccountRow
 import site.remlit.snowdrop.component.Emoji
 import site.remlit.snowdrop.component.ViewSurface
 import site.remlit.snowdrop.model.Account
@@ -36,7 +26,7 @@ import site.remlit.snowdrop.model.ApiResponse
 import site.remlit.snowdrop.model.Emoji
 import site.remlit.snowdrop.model.Reaction
 import site.remlit.snowdrop.util.LocalNavController
-import site.remlit.snowdrop.util.SnackbarController
+import site.remlit.snowdrop.util.LocalSnackbarController
 import site.remlit.snowdrop.util.bgIO
 import snowdrop.shared.generated.resources.Res
 import snowdrop.shared.generated.resources.boosted_by
@@ -54,27 +44,7 @@ fun StatusInteractionDetailView(
 	type: InteractionViewType
 ) = ViewSurface {
 	val navHandler = LocalNavController.current
-	val snackbarHandler = SnackbarController.current
-
-	@Composable
-	fun AccountRow(account: Account) {
-		Row(
-			modifier = Modifier.padding(10.dp),
-			horizontalArrangement = Arrangement.spacedBy(10.dp),
-			verticalAlignment = Alignment.CenterVertically
-		) {
-			Avatar(account)
-
-			Column {
-				Text(account.displayName(), fontWeight = FontWeight.Medium)
-				Text("@${account.acct}")
-			}
-		}
-		HorizontalDivider(
-			thickness = 1.dp,
-			color = MaterialTheme.colorScheme.surfaceContainer
-		)
-	}
+	val snackbarHandler = LocalSnackbarController.current
 
 	TopAppBar(
 		navigationIcon = {